backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1
PyJWT==2.8.0
bcrypt==4.1.2
mysql-connector-python==8.3.0
